pyproject.toml
[tool.poetry]
name = "journey-search"
version = "0.1.0"
description = "Flight Journey Search - FastAPI Challenge"
authors = ["maurofabricioperna@gmail.com"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
pydantic = "^2.8.0"
dependency-injector = "^4.46.0"
python-dotenv = "^1.0.1"
pydantic-settings = "2.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^1.3.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)s %(name)s: %(message)s"
log_cli_date_format = "%H:%M:%S"
